requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==1.10.1
    # via -r requirements.in
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.2
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
cachetools==6.2.1
    # via streamlit
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via streamlit
distro==1.9.0
    # via openai
ebooklib==0.19
    # via -r requirements.in
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
hf-xet==1.1.10
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
huggingface-hub==0.35.3
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
jiter==0.11.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==0.3.27
    # via -r requirements.in
langchain-core==0.3.79
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.11
    # via langchain
langsmith==0.4.34
    # via
    #   langchain
    #   langchain-core
lxml==6.0.2
    # via
    #   ebooklib
    #   python-docx
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
narwhals==2.7.0
    # via altair
networkx==3.5
    # via torch
numpy==2.3.3
    # via
    #   accelerate
    #   pandas
    #   pydeck
    #   streamlit
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.3
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.8.90
    # via torch
openai==2.3.0
    # via -r requirements.in
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   accelerate
    #   altair
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   streamlit
    #   transformers
pandas==2.3.3
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.1
    # via streamlit
psutil==7.1.0
    # via accelerate
pyarrow==21.0.0
    # via streamlit
pydantic==2.12.0
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.41.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pypdf==6.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   accelerate
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.9.18
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
safetensors==0.6.2
    # via
    #   accelerate
    #   transformers
sentencepiece==0.2.1
    # via -r requirements.in
six==1.17.0
    # via
    #   ebooklib
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via langchain
streamlit==1.50.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
tokenizers==0.22.1
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.8.0
    # via
    #   -r requirements.in
    #   accelerate
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.57.0
    # via -r requirements.in
triton==3.4.0
    # via torch
typing-extensions==4.15.0
    # via
    #   altair
    #   beautifulsoup4
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
zstandard==0.25.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# setuptools
